requirements.txt
streamlit_extras
plotly_express
pandas
numpy